requirements.txt
# PostgreSql requirements
psycopg[binary]

# S3 requirements
boto3
botocore

python-dotenv

requests

pytest
pytest-cov

setuptools
wheel
twine